update native loading code a little
updates the getArchString function to support i386, and also support
theoretically any other arch. This won't make oc2r support any arch on
it's own, as we still only ship x86_64/arm64 natives for windows mac and
linux, but it would make it easy for someone to compile their own
natives and add them to the jar.