explicitly set the version of oc2r-native-networking to download
We plan to change the ABI of the native library soon, and if the build
can only download the latest version, it will be impossible to build
commits before the change once the release goes through.  This will mess
with anymore maintaining a fork based on an older version.

Updates to the native networking module will probobly be extremely
uncommon anyway, so I don't see having to manually bump the version
being a major hassle.